bug fix in sequences view and UI improvements
the list of sequences displayed is now dependent on the schema. so its query is now stored as that that needs translation

all navigation introduced by the sidebar keeps the view information and uses it to for the next view. formally the new view information was hard coded into the view elements