bug fix: used start_value in sequences rpr query
in some rare cases. the start_value would differ from the min_value. or the start value would be the max_value if the sequence is descending. The query is now updated to use the start_value insteand of the min_value which might not be the start_value in some other cases (though this cases are not default cases)