second draft: column forms for MySQL dialect
disabled column description in view 'browse' of section 'tbl': I think its not very visually appealing

first draft of creation of generation of sql statements for column creation in 'qry.py'

bug fix: the js libs in the template for the login page hasn't been updated to 1.4 (tt_login.html)

styling of column form for MySQL dialect and fix for the inline script (it rearranges the options field to one line) that was broken

added one more item to the pagination list in the occassion when you are viewing the last page. it usually contain only two anchors with numbers as text. it can now show three of such anchors

added a select_requires form_validation in core.js